api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
boto3==1.42.59
botocore==1.42.59
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
cssselect2==0.9.0
deprecation==2.1.0
fastapi==0.134.0
freetype-py==2.5.1
fsspec==2026.2.0
h11==0.16.0
google-genai==1.34.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jmespath==1.1.0
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
oscrypto==1.3.0
packaging==26.0
pillow==12.1.1
postgrest==2.28.0
propcache==0.4.1
pycairo==1.29.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.2
pypdf==6.7.4
pyroaring==1.0.3
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.28.0
reportlab==4.4.10
requests==2.32.5
rich==14.3.3
rlPyCairo==0.4.0
s3transfer==0.16.0
six==1.17.0
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
svglib==1.6.0
tenacity==9.1.4
tinycss2==1.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
uvicorn==0.41.0
webencodings==0.5.1
websockets==15.0.1
xhtml2pdf==0.2.17
yarl==1.22.0
zstandard==0.25.0
